# OpenApi.yml
openapi: '3.0.3'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://localhost:3000

components:
  schemas:
    user:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        age:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string

paths:
  /users:
    get:
      description: 'endpoint untuk mendapatkan semua data user'
      parameters: []
      responses:
        '200':
          description: Berhasil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/user'
        '500':
          description: Server Gagal
        '403':
          description: Tidak diizinkan
    
    post:
      description: 'Menambahkan user baru'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: Berhasil menambahkan user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Data tidak lengkap
        '500':
          description: Server Gagal

  /users/{id}:
    get:
      description: 'endpoint untuk mendapatkan data user berdasarkan id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
          '200':
            description: Berhasil
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/user'
          '500':
            description: Server Gagal
          '403':
            description: Tidak diizinkan

    delete:
      description: 'endpoint untuk menghapus data user berdasarkan id'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
          '200':
            description: Berhasil
          '500':
            description: Server Gagal
          '403':
            description: Tidak diizinkan
    put:
      description: 'Endpoint untuk mengubah data user berdasarkan ID'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                age:
                  type: integer
      responses:
        '200':
          description: Berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '400':
          description: Request body tidak lengkap
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Server Gagal
        